+++ src/components/draw/CanvasPalette.js
@@ -1,4 +1,5 @@
 import React, { useState, useRef, useEffect } from "react";
+import BrushSizeControl from "./BrushSizeControl";
 import DrawingToolBar from "./DrawingToolBar";
 import { TbTriangleFilled, TbTriangleInvertedFilled } from "react-icons/tb";
 
@@ -48,32 +49,6 @@ const PaletteToggle = ({ isOpen, togglePalette }) => {
       className="bg-[#d9d9d9] flex justify-center items-center py-2 rounded-ss-2xl rounded-se-2xl cursor-pointer"
     >
       {isOpen ? <TbTriangleInvertedFilled /> : <TbTriangleFilled />}
-    </div>
-  );
-};
-
-const BrushSizeControl = ({ selectedColor }) => {
-  const [circleSize, setCircleSize] = useState(20); // 초기 원의 크기 설정
-  return (
-    <div className="flex items-center mx-3">
-      <input
-        type="range"
-        min="20"
-        max="50"
-        value={circleSize}
-        onChange={(event) => setCircleSize(parseInt(event.target.value))}
-        className="mr-4 flex-1"
-      />
-      <div className="w-16 h-16 flex justify-center items-center">
-        <div
-          className={`rounded-full ${selectedColor === "transparent" ? "border-4" : ""}`}
-          style={{
-            width: `${circleSize}px`,
-            height: `${circleSize}px`,
-            backgroundColor: selectedColor,
-          }}
-        />
-      </div>
     </div>
   );
 };
